DentalDiseasesClassification/requirements.txt
Flask
Flask-CORS
torch
torchvision
pillow
numpy
